pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-stock-exchange"
version = "0.1.0"
description = "OpenEnv Stock Exchange RL Environment"
readme = "README.md"
requires-python = ">=3.9"
authors = [
    { name = "Ashmit Sahu" }
]
dependencies = [
    "fastapi",
    "uvicorn",
    "websockets",
    "pydantic",
    "openai",
    "httpx",
    "python-dotenv",
    "gymnasium",
    "numpy",
    "jinja2",
    "python-multipart",
    "openenv-core"
]

[project.optional-dependencies]
rl = [
    "stable-baselines3"
]

[project.urls]
"Homepage" = "https://github.com/ashmitsahu1812/ShieldX-Auto-DPO"

[project.scripts]
server = "server.app:main"

[project.entry-points."console_scripts"]
server = "server.app:main"

[tool.setuptools]
packages = ["server", "rl"]

[tool.setuptools.package-data]
"*" = ["static/*", "static/**/*"]
